fix(go): json un/marshaling for AnyOf types
Without custom json marshaling methods,
AnyOf types are marshaled incorrectly to json
because they include all of the anyof fields.

This solution assumes that the AnyOf field
types will be pointers (and can be checked if nil),
is assumes that there are TWO AnyOf types,
and that they are (ordered) a singleton of
Any type, AND a [] of Any type.

Default json un/marshaling will be used
for the first single existing field type.

Signed-off-by: meows <b5c6@protonmail.com>