fix(go): the naive way of un/marshaling macros
This is naive because the marshaling methods
assume we want to maintain the array data
structure inherited from the JSON schema definition.
BUT, a OneOfStuff type should not marshal and unmarshal
to and from an array, but from a single struct --
because it's _One_Of.
The point is not to maintain the structure of the
_definition_, but to deduce the logic behavior
of the definition as embodied in an object.

Signed-off-by: meows <b5c6@protonmail.com>