fix(go): implement AnyOf and OneOf un/marshaling methods
These json schema types describe composite objects.
The typed objects should never return or accept
(incidentally validate) json arrays.
The un/marshaling logic is implemented as sensibly as I
could, but there are nuances that are not handled,
eg. should OneOf unmarshaling throw an error if the
the object succeeds in unmarshaling more than one
of its composing properties? As is, does not throw,
just returns on first ok.

Signed-off-by: meows <b5c6@protonmail.com>