fix(integration-tests): readable fake-object field names
Rename field names from very whimsical
to slightly less whimsical.

Updates fake-object field names and their
codegen target.json counterparts (stub and test control).

This is a combination of 2 commits.

This is the 1st commit message:

This is a combination of 6 commits.

This is the 1st commit message:

fix(integration-tests): property names for readability, moebius

Signed-off-by: meows <b5c6@protonmail.com>

This is the commit message #2:

fix(integration-tests): 'complex' names, kitchen sink

Updates tests

This is the commit message #3:

fix(integration-tests): duplicate-names.json

Signed-off-by: meows <b5c6@protonmail.com>

This is the commit message #4:

fix(integration-tests): multi-schema

Signed-off-by: meows <b5c6@protonmail.com>

This is the commit message #5:

fix(integration-tests): pets.json

Signed-off-by: meows <b5c6@protonmail.com>

This is the commit message #6:

remove file write

Signed-off-by: meows <b5c6@protonmail.com>

This is the commit message #2:

autoformat

Signed-off-by: meows <b5c6@protonmail.com>